le colocamos seguridad a las rutas y agregamos el poder colocar o cambiar roles desde usuarios

+++ app/Http/Controllers/UserController.php
@@ -5,16 +5,20 @@ namespace App\Http\Controllers;
 use App\Http\Requests\UserRequest;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+use Spatie\Permission\Models\Role;
 
 class UserController extends Controller
 {
+    use AuthorizesRequests;
     /**
      * Display a listing of the resource.
      */
     public function index(Request $request)
     {
+        $this->authorize('user-list');
         $texto = $request->input("texto");
-         $registros = User::where("name",'like','%'.$texto.'%')
+         $registros = User::with('roles')->where('name','like','%'.$texto.'%')
          ->orWhere('email','like','%'.$texto.'%')
          ->orderBy('id','desc')
          ->paginate(10);
@@ -26,7 +30,9 @@ class UserController extends Controller
      */
     public function create()
     {
-        return view('usuario.action');
+        $this->authorize('user-create');
+        $roles=Role::all();
+        return view('usuario.action', compact('roles'));
     }
 
     /**
@@ -34,12 +40,15 @@ class UserController extends Controller
      */
     public function store(UserRequest $request)
     {
+        $this->authorize('user-edit');
         $registro=new User();
         $registro->name=$request->input('name');
         $registro->email=$request->input('email');
         $registro->password=bcrypt($request->input('password'));
         $registro->activo=$request->input('activo');
         $registro->save();
+
+        $registro->assignRole($request->input('role'));
         return redirect()->route('usuarios.index')->with('mensaje', 'registro '.$registro->name. ' Agregado correctamente');
     }
 
@@ -56,9 +65,11 @@ class UserController extends Controller
      */
     public function edit($id)
     {
+        $this->authorize('user-edit');
+        $roles=Role::all();
         $registro = User::findOrFail($id);
 
-        return view('usuario.action', compact('registro'));
+        return view('usuario.action', compact('registro', 'roles'));
     }
 
     /**
@@ -66,12 +77,15 @@ class UserController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $this->authorize('user-edit');
         $registro=User::findOrFail($id);
         $registro->name=$request->input('name');
         $registro->email=$request->input('email');
         $registro->password=bcrypt($request->input('password'));
         $registro->activo=$request->input('activo');
         $registro->save();
+
+        $registro->assignRole($request->input('role'));
         return redirect()->route('usuarios.index')->with('mensaje', 'El Registro de '.$registro->name.' actualizado satisfactoriamente');
     }
 
@@ -80,12 +94,14 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('user-delete');
         $registro=User::findOrFail($id);
         $registro->delete();
         return redirect()->route('usuarios.index')->with('mensaje', $registro->name. ' <Eliminado Satisfactoriamente.');
     }
 
     public function toggleStatus(User $usuario){
+        $this->authorize('user-activate');
         $usuario->activo =!$usuario->activo;
         $usuario->save();
         return redirect()->route('usuarios.index')->with('mensaje', 'El estado del usuario ha sido actualizado satisfactoriamente');
